requirements.txt
discord.py~=1.6.0
python-dotenv~=0.15.0
pip~=20.0.2
wheel~=0.34.2
cryptography~=2.8
keyring~=18.0.1
setuptools~=45.2.0